Refactoring code with better folder structure
