+++ docker-compose.yml
@@ -35,7 +35,7 @@ services:
       sh -c "
         cd /app &&
         python -m dev_bootstrap.bootstrap &&
-        uvicorn api.main:app --host 0.0.0.0 --port 8000
+        uvicorn app.main:app --host 0.0.0.0 --port 8000
       "
 
   db:
