+++ python/app/api/dependencies.py
@@ -1,0 +1,8 @@
+from app.db.session import SessionLocal
+
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
