+++ python/app/config.py
@@ -1,0 +1,3 @@
+
+# replace db with localhost if you are spinning up fast api in local instead of docker
+DATABASE_URL = "postgresql://indimitra:indimitra123@db:5432/indimitra"
