+++ python/app/graphql/schema.py
@@ -1,0 +1,4 @@
+import strawberry
+from app.graphql.resolvers.product_resolver import Query, Mutation
+
+schema = strawberry.Schema(query=Query, mutation=Mutation)
